// bookshop_frontend/src/app/model/order/order.ts
export class Order {
  orderId: number | undefined;
  orderNumber: number | undefined;
  shopId: number | undefined;
  cost: number | undefined;
  deliveryAddress: string | undefined;
  description: string | undefined;
  orderSubmitDate: string | undefined;
  classification: string | undefined;
  orderCompleteDate: string | undefined;
  username: string | undefined;
}